Refactor AES encryption tests and methods
The commit refactors AES encryption tests and methods to use DigestAlgorithm instead of the previously used 'Algorithm' inner class in AES. The changes also improve code consistency by adding 'this' keyword to object property accesses and reformatting some blocks for readability. There are also some changes to the import statements to adapt to the new changes.